Add placeholder live site domain
Now all the request should be done through the api gateway `api`
route, which should look like this `https://<the domain>/api`.
This change takes in account the development environment which
runs on the local environment with a fixed Class C IP address.

The most relevant change in the package.json file is the `homepage`
field which is required by the build scripts, the other ones are
just grooming changes.